+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityZombie.java
@@ -1,7 +1,8 @@
 package org.allaymc.api.entity.interfaces;
 
+import org.allaymc.api.entity.component.EntityContainerHolderComponent;
 import org.allaymc.api.entity.component.EntityHeadYawComponent;
 import org.allaymc.api.entity.component.EntityUndeadComponent;
 
-public interface EntityZombie extends EntityIntelligent, EntityHeadYawComponent, EntityUndeadComponent {
+public interface EntityZombie extends EntityIntelligent, EntityHeadYawComponent, EntityUndeadComponent, EntityContainerHolderComponent {
 }