+++ server/src/main/java/org/allaymc/server/container/impl/PlayerArmorContainerImpl.java
@@ -7,17 +7,19 @@ import org.allaymc.api.item.ItemHelper;
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.item.data.ArmorTier;
 import org.allaymc.api.item.interfaces.ItemAirStack;
+import org.cloudburstmc.nbt.NbtMap;
 
+import java.util.List;
 import java.util.function.Supplier;
 
 /**
  * @author daoge_cmd
  */
-public class ArmorContainerImpl extends AbstractPlayerContainer implements ArmorContainer {
+public class PlayerArmorContainerImpl extends AbstractPlayerContainer implements ArmorContainer {
 
     protected static final float KNOCKBACK_RESISTANCE_PER_NETHERITE_ARMOR = 0.1f;
 
-    public ArmorContainerImpl(Supplier<EntityPlayer> playerSupplier) {
+    public PlayerArmorContainerImpl(Supplier<EntityPlayer> playerSupplier) {
         super(ContainerTypes.ARMOR, playerSupplier);
         addSlotChangeListener(0, this::onArmorChange);
         addSlotChangeListener(1, this::onArmorChange);
@@ -25,8 +27,7 @@ public class ArmorContainerImpl extends AbstractPlayerContainer implements Armor
         addSlotChangeListener(3, this::onArmorChange);
     }
 
-    protected void onArmorChange(ItemStack newItemStack) {
-        // Recalculate knockback resistance
+    protected void onArmorChange(ItemStack $) {
         var knockbackResistance = 0.0f;
         for (var itemStack : content) {
             if (itemStack == ItemAirStack.AIR_STACK ||
@@ -36,10 +37,16 @@ public class ArmorContainerImpl extends AbstractPlayerContainer implements Armor
 
             knockbackResistance += KNOCKBACK_RESISTANCE_PER_NETHERITE_ARMOR;
         }
+
         var player = playerSupplier.get();
         player.setKnockbackResistance(knockbackResistance);
-        // Send armor to viewers
         player.forEachViewers(viewer -> viewer.viewEntityArmors(player));
+    }
+
+    @Override
+    public void loadNBT(List<NbtMap> nbtList) {
+        super.loadNBT(nbtList);
+        onArmorChange(null);
     }
 
     @Override